django-version/requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/dev.in
#
asgiref==3.8.1
    # via
    #   -r requirements/prod.txt
    #   django
    #   django-cors-headers
attrs==24.2.0
    # via
    #   -r requirements/prod.txt
    #   jsonschema
    #   referencing
black==24.10.0
    # via -r requirements/dev.in
build==1.2.2.post1
    # via pip-tools
certifi==2024.8.30
    # via
    #   -r requirements/prod.txt
    #   requests
cffi==1.17.1
    # via
    #   -r requirements/prod.txt
    #   cryptography
charset-normalizer==3.4.0
    # via
    #   -r requirements/prod.txt
    #   requests
click==8.1.7
    # via
    #   black
    #   pip-tools
coverage==7.6.4
    # via -r requirements/dev.in
cryptography==43.0.3
    # via -r requirements/prod.txt
django==5.1.2
    # via
    #   -r requirements/prod.txt
    #   django-cors-headers
    #   django-extensions
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-cors-headers==4.5.0
    # via -r requirements/prod.txt
django-extensions==3.2.3
    # via -r requirements/prod.txt
djangorestframework==3.15.2
    # via
    #   -r requirements/prod.txt
    #   drf-spectacular
drf-spectacular==0.27.2
    # via -r requirements/prod.txt
drf-spectacular-sidecar==2024.7.1
    # via -r requirements/prod.txt
factory-boy==3.3.1
    # via -r requirements/dev.in
faker==30.8.0
    # via factory-boy
flake8==7.1.1
    # via -r requirements/dev.in
freezegun==1.5.1
    # via -r requirements/dev.in
gevent==24.10.3
    # via
    #   -r requirements/dev.in
    #   -r requirements/prod.txt
greenlet==3.1.1
    # via
    #   -r requirements/prod.txt
    #   gevent
gunicorn==23.0.0
    # via -r requirements/prod.txt
hvac==2.3.0
    # via -r requirements/prod.txt
idna==3.10
    # via
    #   -r requirements/prod.txt
    #   requests
inflection==0.5.1
    # via
    #   -r requirements/prod.txt
    #   drf-spectacular
jsonschema==4.23.0
    # via
    #   -r requirements/prod.txt
    #   drf-spectacular
jsonschema-specifications==2024.10.1
    # via
    #   -r requirements/prod.txt
    #   jsonschema
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==24.1
    # via
    #   -r requirements/prod.txt
    #   black
    #   build
    #   gunicorn
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via -r requirements/dev.in
platformdirs==4.3.6
    # via black
pycodestyle==2.12.1
    # via flake8
pycparser==2.22
    # via
    #   -r requirements/prod.txt
    #   cffi
pyflakes==3.2.0
    # via flake8
pymysql==1.1.1
    # via
    #   -r requirements/dev.in
    #   -r requirements/prod.txt
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via
    #   faker
    #   freezegun
python-decouple==3.8
    # via -r requirements/prod.txt
pyyaml==6.0.2
    # via
    #   -r requirements/prod.txt
    #   drf-spectacular
referencing==0.35.1
    # via
    #   -r requirements/prod.txt
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements/prod.txt
    #   hvac
rpds-py==0.20.0
    # via
    #   -r requirements/prod.txt
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
sqlparse==0.5.1
    # via
    #   -r requirements/prod.txt
    #   django
typing-extensions==4.12.2
    # via faker
uritemplate==4.1.1
    # via
    #   -r requirements/prod.txt
    #   drf-spectacular
urllib3==2.2.3
    # via
    #   -r requirements/prod.txt
    #   requests
wheel==0.44.0
    # via pip-tools
zope-event==5.0
    # via
    #   -r requirements/prod.txt
    #   gevent
zope-interface==7.1.1
    # via
    #   -r requirements/prod.txt
    #   gevent

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
